power panic: Move code into a separate file
