+++ Firmware/power_panic.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void uvlo_();
+void recover_print(uint8_t automatic); 
+void setup_uvlo_interrupt();
